Store created teacher in persons.json file
