+++ storage/persons.json
@@ -1,1 +1,1 @@
-[{"json_class":"Student","id":820,"age":56,"name":"basit","parent_permission":true},{"json_class":"Student","id":566,"age":57,"name":"admin","parent_permission":true}]
+[{"json_class":"Student","id":820,"age":56,"name":"basit","parent_permission":true},{"json_class":"Student","id":566,"age":57,"name":"admin","parent_permission":true},{"json_class":"Teacher","id":192,"age":34,"name":"Alphayo","specialization":"Physicist"}]
