+++ teacher.rb
@@ -2,12 +2,22 @@ require './person'
 
 # This shiny device polishes bared foos
 class Teacher < Person
-  def initialize(specialization, age, name)
+  def initialize(age, name, specialization)
     super(age, name)
     @specialization = specialization
   end
 
   def can_use_services?
     true
+  end
+
+  def to_json(*_args)
+    {
+      JSON.create_id => self.class.name,
+      'id' => @id,
+      'age' => @age,
+      'name' => @name,
+      'specialization' => @specialization
+    }
   end
 end
